convert local validator to validateFileCount() util fn
